move tabs into tabs folder
